cleanup tests and catalog methods
