+++ src/models/Meeting.ts
@@ -4,11 +4,6 @@
  */
 
 /**
- *
- */
-import { RawMeeting } from './Meta';
-
-/**
  * A meeting represents a specific meeting time \w information about the instructor and location
  *
  * It has a reference back to the section that it belongs to
@@ -25,16 +20,7 @@ export default class Meeting {
         return profs;
     }
 
-    public readonly instructor: string;
-    public readonly days: string;
-    public readonly room: string;
-    constructor(raw: RawMeeting) {
-        this.instructor = raw[0];
-        this.days = raw[1];
-        this.room = raw[2];
-    }
-
-    public sameTimeAs(other: Meeting) {
-        return this.days === other.days;
-    }
+    public readonly instructor!: string;
+    public readonly days!: string;
+    public readonly room!: string;
 }
